pyproject.toml
[project]
name = "python-project-52"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "django-bootstrap5>=25.3",
    "django-filter>=25.2",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.14.10",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]
